Berger walking forward images
These are the images I have so far of the berger walking forward. I wanted to give the images as soon as possible and you can start working with these however currently I am working on making the design more realistic given that currently only the legs move not the rest of the body or face.